7.1.4 Refactor reducer, split for 3 own reducers.
