+++ src/store/action.js
@@ -7,15 +7,15 @@ export const ActionType = {
   LOAD_COMMENTS: `data/loadComments`,
   SET_DATA: `data/setData`,
   SET_ACTIVE: `data/setActive`,
-  CHANGE_GENRE: `films/changeGenre`,
-  GET_MOVIES: `films/getMovies`,
-  GET_MY_MOVIES: `films/getMyMovies`,
-  RESET_APP: `films/reset`,
-  RESET_ACTIVE_MOVIE: `films/resetActiveMovie`,
-  CHANGE_CONTENT: `films/changeContent`,
-  INCREMENT_STEP: `films/incrementStep`,
-  REDIRECT_TO_ROUTE: `films/redirectToRoute`,
-  SET_CONTENT_REVIEW: `films/setContentReview`,
+  CHANGE_GENRE: `movies/changeGenre`,
+  GET_MOVIES: `movies/getMovies`,
+  GET_MY_MOVIES: `movies/getMyMovies`,
+  RESET_APP: `movies/reset`,
+  RESET_ACTIVE_MOVIE: `movies/resetActiveMovie`,
+  CHANGE_CONTENT: `movies/changeContent`,
+  INCREMENT_STEP: `movies/incrementStep`,
+  REDIRECT_TO_ROUTE: `movies/redirectToRoute`,
+  SET_CONTENT_REVIEW: `movies/setContentReview`,
   REQUIRED_AUTHORIZATION: `user/requiredAuthorization`,
   SET_LOGIN_ERROR: `user/setLoginError`,
 };
